feat(org): dedicated Team tab for member roster + RBAC access management
Promotes team/access management out of the buried shield-icon-on-Overview
entry point into its own admin/owner-gated surface.

- New "Team" tab in Organization Settings (Overview · Team · Business · …),
  reachable from the sidebar (Settings → Team), the settings nav rail, and a
  "Manage team" button on the Overview roster.
- Overview keeps a read-only team roster (role badges + access chip, admin
  summary only) — no role editing, remove, or RBAC-editor entry there.
- Team tab holds the full management surface: editable roles, remove member,
  invitations, and an explicit "Manage access" button (replacing the mystery
  shield icon) that opens the existing full-page per-member RBAC editor.
- Shared TeamMembersTable(readOnly) drives both tabs so they never drift.
- Team tab + sidebar item are admin/owner-only (hasFullAccess); members get a
  read-only roster on Overview and a no-access panel if they deep-link to it.
- Member editor "Back" returns to the Team tab.

Extracts shared Clerk role/member helpers to _lib/org-members.ts.
